# openapi/src/main/openapi3/master_rest_v1.yaml
...
openapi: 3.0.0
servers:
  - url: http://localhost:9098/
    description: Localhost reference server
security:
  - basic: []
info:
  title: Celeborn Master REST API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
paths:
...
  /api/v1/masters:
    get:
      tags:
        - Master
      operationId: getMasterGroupInfo
      description: |
        List master group information of the service.
        It will list all master's LEADER, FOLLOWER information.
      responses:
        "200":
          description: The request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterInfoResponse'
        "400":
          description: HA is not enabled.
...
  /api/v1/workers/exclude:
    post:
      tags:
        - Worker
      operationId: excludeWorker
      description: |
        Excluded workers of the master add or remove the worker manually given worker id.
        The parameter add or remove specifies the excluded workers to add or remove.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExcludeWorkerRequest'
      responses:
        "200":
          description: The request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandleResponse'
  /api/v1/workers/events:
    get:
      tags:
        - Worker
      operationId: getWorkerEvents
      description: List all worker event infos of the master.
      responses:
        "200":
          description: The request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerEventData'
    post:
      tags:
        - Worker
      operationId: sendWorkerEvent
      description: |
        For Master(Leader) can send worker event to manager workers.
        Legal types are 'None', 'Immediately', 'Decommission', 'DecommissionThenIdle', 'Graceful', 'Recommission'.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendWorkerEventRequest'
      responses:
        "200":
          description: The request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandleResponse'
...
components:
  schemas:
...
    MasterLeader:
      type: object
      properties:
        id:
          type: string
          description: The id of the master leader raft peer.
        address:
          type: string
          description: The raft RPC address of the master leader.
      required:
        - id
        - address
    MasterCommitData:
      type: object
      properties:
        commitIndex:
          type: integer
          format: int64
          description: The raft commit index of the master.
        id:
          type: string
          description: The id of the raft peer.
        address:
          type: string
          description: The raft RPC address of the peer for server-server communication.
        clientAddress:
          type: string
          description: The raft RPC address of the peer for client operations.
        startUpRole:
          type: string
          description: The raft start up role of the master.
      required:
        - commitIndex
        - id
        - address
        - clientAddress
        - startUpRole
    MasterInfoResponse:
      type: object
      properties:
        groupId:
          type: string
          description: The group id of the master raft server.
        leader:
          $ref: '#/components/schemas/MasterLeader'
        masterCommitInfo:
          type: array
          default: [ ]
          description: The raft commit info of the master.
          items:
            $ref: '#/components/schemas/MasterCommitData'
      required:
        - groupId
    WorkerData:
      type: object
      properties:
        host:
          type: string
          description: The host of the worker.
        rpcPort:
          type: integer
          format: int32
          description: The rpc port of the worker.
        pushPort:
          type: integer
          format: int32
          description: The push port of the worker.
        fetchPort:
          type: integer
          format: int32
          description: The fetch port of the worker.
        replicatePort:
          type: integer
          format: int32
          description: The replicate port of the worker.
        internalPort:
          type: integer
          format: int32
          description: The internal port of the worker.
        slotUsed:
          type: integer
          format: int64
          description: The slot used of the worker.
        lastHeartbeatTimestamp:
          type: integer
          format: int64
          description: The last heartbeat timestamp of the worker.
        heartbeatElapsedSeconds:
          type: integer
          format: int64
          description: The elapsed seconds since the last heartbeat of the worker.
        diskInfos:
          type: object
          description: A map of disk name and disk info.
          additionalProperties:
            type: string
        resourceConsumption:
          type: object
          description: A map of identifier and resource consumption.
          additionalProperties:
            type: string
        workerRef:
          type: string
          description: The reference of the worker.
        workerState:
          type: string
          description: The state of the worker.
        workerStateStartTime:
          type: integer
          format: int64
          description: The start time of the worker state.
      required:
        - host
        - rpcPort
        - pushPort
        - fetchPort
        - replicatePort
...
    WorkerEventInfoData:
      type: object
      properties:
        eventType:
          type: string
          description: The type of the event.
        eventTime:
          type: integer
          format: int64
          description: The start time of the event.
      required:
        - eventType
        - eventTime
    WorkerEventData:
      type: object
      properties:
        worker:
          $ref: '#/components/schemas/WorkerData'
        event:
          $ref: '#/components/schemas/WorkerEventInfoData'
      required:
        - worker
        - event
...
    WorkerId:
      type: object
      properties:
        host:
          type: string
          description: The host of the worker.
        rpcPort:
          type: integer
          format: int32
          description: The rpc port of the worker.
        pushPort:
          type: integer
          format: int32
          description: The push port of the worker.
        fetchPort:
          type: integer
          format: int32
          description: The fetch port of the worker.
        replicatePort:
          type: integer
          format: int32
          description: The replicate port of the worker.
      required:
        - host
        - rpcPort
        - pushPort
        - fetchPort
        - replicatePort
    ExcludeWorkerRequest:
      type: object
      properties:
        add:
          type: array
          default: [ ]
          description: The workers to be added to the excluded workers.
          items:
            $ref: '#/components/schemas/WorkerId'
        remove:
          type: array
          default: [ ]
          description: The workers to be removed from the excluded workers.
          items:
            $ref: '#/components/schemas/WorkerId'
    SendWorkerEventRequest:
      type: object
      properties:
        eventType:
          type: string
          default: None
          description: The type of the event.
          enum:
            - Immediately
            - Decommission
            - DecommissionThenIdle
            - Graceful
            - Recommission
            - None
        workers:
          type: array
          default: [ ]
          description: The workers to send the event.
          items:
            $ref: '#/components/schemas/WorkerId'
      required:
        - eventType
        - worker
    HandleResponse:
      type: object
      properties:
        success:
          type: boolean
          description: The result of the request.
        message:
          type: string
          description: The message of the request.
      required:
        - success
  securitySchemes:
    basic:
      type: http
      scheme: basic